fixed most bugs, still a bit of hack since vosk seems like a hack of library that is really just Kaldi with fancy shoes. looks like linker issues with Processor::Verbalize

+++ src/apps/file_to_vosk_to_file.c
@@ -59,7 +59,7 @@ int main(int argc, char *argv[])
         p_read_file = strdup(optarg);
         break;
       case 's':
-        rate = atof(optarg);
+        rate = (float)atof(optarg);
         break;
       case 'h':
       default:
